// football-app-client/src/api/models/HighlightDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type HighlightDto = {
    url: string;
    homeTeam: string;
    awayTeam: string;
    competition: string;
    matchDate: string;
    id: string;
    uploadDate: string;
    likes: number;
    thumbnailUrl: string;
};